Dynamic sound adaption
Clip will now be louder then the song itself to standout.
This works by getting the duration, then taking a randomint of that duration, getting the decibel, adding the decibel amount of the clip to the clip and then apply_gaining.

I also cleaned up the dependancies, since those were really messy.

Testing is not done yet, this gain method is not perfect yet.